kunit: Add kunit_tool support in kernelci-core
--kunit and --kunit_json flags have been added that can be specified
with the "./kci_build build_kernel" command.  If --kunit is specified by
the user, the script will first check if kunit is present at
<kdir>/tools/testing/kunit/kunit.py.  Further, --kunit needs to be set
to a particular value from [alltests, defconfig, kunitconfig] (i.e.
--kunit=kunitconfig), where this value specifies whether to run
alltests, defconfig kunit tests or tests that can be configured with a
user-created .kunitconfig file. If kunit_tool is present, the kunit.py
run command is executed and the parsed test results are printed to
console as well as stored as a JSON in a file specified by --kunit_json.

Signed-off-by: Heidi Fahim <heidifahim@google.com>
Change-Id: I321a8c95438cfede3cceffe18e43c6e039981e15